ai-agent-resell-guide/requirements.txt
streamlit==1.28.1
langchain==0.0.350
langchain-openai==0.0.2
langchain-community==0.0.10
chromadb==0.4.18
openai==1.3.7
python-dotenv==1.0.0
docx2txt==0.8
pypdf==3.17.1
tiktoken==0.5.1 